requirements.txt
python-dotenv
pyautogui
pyautogui pynput
